requirements.txt
asgiref==3.2.10
Babel==2.8.0
Django==3.0.7
phonenumbers==8.12.6
Pillow==7.1.2
psycopg2-binary==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1